rename auth api
