+++ src/api/index.js
@@ -1,1 +1,1 @@
-export * from './auth';
+export * from './user';
